Refactoring in ASR code on iOS
