+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Refactoring in ASR code on macOS
+- Refactoring in ASR code on iOS
 
 ## [0.8.353] - 2023-05-18
 ### Added:
